MM-8589: expunge global mm_config from system analytics component
